+++ src/js/program.js
@@ -2,7 +2,7 @@
 
                   This file is part of the Buildbotics firmware.
 
-         Copyright (c) 2015 - 2023, Buildbotics LLC, All rights reserved.
+         Copyright (c) 2015 - 2026, Buildbotics LLC, All rights reserved.
 
           This Source describes Open Hardware and is licensed under the
                                   CERN-OHL-S v2.
@@ -31,15 +31,33 @@ let util = require('./util')
 
 class Program {
   constructor($api, path) {
-    this.$api = $api
-    this.path = path
-    this.filename = util.display_path(path)
-    this.progress = 0
+    this.$api      = $api
+    this.path      = path
+    this.filename  = util.display_path(path)
+    this.progress  = 0
+    this.timestamp = Date.now()
+  }
+
+
+  // Invalidate cached data so files are re-fetched from server
+  invalidate() {
+    this._load      = null
+    this._toolpath  = null
+    this._positions = null
+    this._speeds    = null
+    this._view      = null
+    this.timestamp  = Date.now()
+  }
+
+
+  async download(url, conf = {}) {
+    conf.params = (conf.params || {})['_t']= this.timestamp // cache-busting
+    return this.$api.get(url, conf)
   }
 
 
   async download_array(url) {
-    let data = await this.$api.get(url, {type: 'arraybuffer'})
+    let data = await this.download(url, {type: 'arraybuffer'})
     return new Float32Array(data)
   }
 
@@ -76,7 +94,7 @@ class Program {
 
 
   async _load_toolpath() {
-    let toolpath = await this.$api.get('path/' + this.path)
+    let toolpath = await this.download('path/' + this.path)
 
     if (toolpath.progress == undefined) {
       this.progress = 1
@@ -98,7 +116,8 @@ class Program {
 
   load() {
     if (!this._load)
-      this._load = this.$api.get('fs/' + this.path, {type: 'text'})
+      this._load = this.download('fs/' + this.path, {type: 'text'})
+
     return this._load
   }
 }